+++ starfish/experiment/builder.py
@@ -1,7 +1,7 @@
-from starfish.core.experiment.builder import (   # noqa: F401
+from starfish.core.experiment.builder import inplace  # noqa: F401
+from starfish.core.experiment.builder.builder import (   # noqa: F401
     build_image,
     build_irregular_image,
-    inplace,
     TileIdentifier,
     write_experiment_json,
     write_irregular_experiment_json,
